Make NoteText view more generic.
